requirements.txt
openai>=1.93.0
requests>=2.32.3
pydantic>=2.11.5
python-dotenv>=1.1.0
tavily-python>=0.7.4
fastapi>=0.114.12
uvicorn>=0.34.3
typing-extensions>=4.13.2
gradio>=5.37.0
markdown>=3.8.2 